refactor to creatre sprint card
