+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@atoll/shared",
-  "version": "0.19.0",
+  "version": "0.20.0",
   "author": {
     "name": "Kevin Berry",
     "email": "41717340+51ngul4r1ty@users.noreply.github.com"
